- Implemented adding a new category in the menu - Removed adding of the items button from Item availability - Working good in localhost
